refactor: 코어 계층 토큰 정리 로직 분리
